Moved the message processing to its own channel
