+++ plugin/src/main/java/at/helpch/chatchat/util/MessageProcessor.java
@@ -1,0 +1,58 @@
+package at.helpch.chatchat.util;
+
+import at.helpch.chatchat.ChatChatPlugin;
+import at.helpch.chatchat.api.Channel;
+import at.helpch.chatchat.api.ChatUser;
+import at.helpch.chatchat.api.event.ChatChatEvent;
+import net.kyori.adventure.audience.Audience;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.format.NamedTextColor;
+import org.jetbrains.annotations.NotNull;
+
+public class MessageProcessor {
+    private static final String UTF_PERMISSION = "chatchat.utf";
+
+    public static void process(
+        @NotNull final ChatChatPlugin plugin,
+        @NotNull final ChatUser user,
+        @NotNull final Channel channel,
+        @NotNull final String message,
+        final boolean async
+    ) {
+        if (StringUtils.containsIllegalChars(message) && !user.player().hasPermission(UTF_PERMISSION)) {
+            user.sendMessage(Component.text("You can't use special characters in chat!", NamedTextColor.RED));
+            return;
+        }
+
+        final var format = FormatUtils.findFormat(user.player(), plugin.configManager().formats());
+
+        final var audience = plugin.usersHolder().users()
+            .stream()
+            .filter(otherUser -> otherUser.canSee(channel)) // get everyone who can see this channel
+            .collect(Audience.toAudience());
+
+        final var chatEvent = new ChatChatEvent(
+            async,
+            user,
+            audience,
+            format,
+            Component.text(message),
+            channel
+        );
+
+        plugin.getServer().getPluginManager().callEvent(chatEvent);
+
+        if (chatEvent.isCancelled()) {
+            return;
+        }
+
+        final var oldChannel = user.channel();
+        user.channel(channel);
+        chatEvent.recipients().sendMessage(FormatUtils.parseFormat(
+            chatEvent.format(),
+            user.player(),
+            chatEvent.message()
+        ));
+        user.channel(oldChannel);
+    }
+}
